Add files via upload
A string is an array of characters.

In this question, the string is a list of characters that ends by pressing the enter button.
In order to indicate that a string is finished, the zero skid code ('0\') is placed in the end house.
This character, which is used specifically for this purpose, is called nullCharacter.
Due to its high usage, this character is automatically added to the string when it is received, and the programmer does not need to do anything special.
Well, now we turn to the question:

You have to write a program that first receives a string and two characters (a,b) from the user. Then, in the output, first print the length of the input string, then display the reverse of the substring of the entered string that starts with the character a and ends with the character b. (Pay attention to the examples)

Obviously, a, b must be present in the input string.
It should be noted that using the strlen function to display the length of the input string is not allowed.